+++ tests/testthat/helper-vdiffr.R
@@ -1,0 +1,2 @@
+
+theme_set(theme_test())