+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * left- or right-aligned facet labels with multiple lines are now correctly
   aligned. In previous versions, the lines of the text would still be centered.
+  (@mkuhn, #2023)
+
+* Automatic visual unit tests with vdiffr.
 
 * Layers no longer warn about unknown aesthetics who's value is set to 
   `NULL` (overriding a set aesthetic in the plot) (#1909).